Kernel - Update VFS API to use off_t/size_t instead of Uint64
- Also fixed compilation issues in x86_64/ARMv7

+++ KernelLand/Kernel/drv/proc.c
@@ -33,7 +33,7 @@ typedef struct sSysFS_Ent
 
 char	*SysFS_Comm_ReadDir(tVFS_Node *Node, int Id);
 tVFS_Node	*SysFS_Comm_FindDir(tVFS_Node *Node, const char *Filename);
-Uint64	SysFS_Comm_ReadFile(tVFS_Node *Node, Uint64 Offset, Uint64 Length, void *Buffer);
+size_t	SysFS_Comm_ReadFile(tVFS_Node *Node, off_t Offset, size_t Length, void *Buffer);
 void	SysFS_Comm_CloseFile(tVFS_Node *Node);
 
 // === GLOBALS ===
@@ -357,10 +357,9 @@ tVFS_Node *SysFS_Comm_FindDir(tVFS_Node *Node, const char *Filename)
 }
 
 /**
- * \fn Uint64 SysFS_Comm_ReadFile(tVFS_Node *Node, Uint64 Offset, Uint64 Length, void *Buffer)
  * \brief Read from an exposed buffer
  */
-Uint64 SysFS_Comm_ReadFile(tVFS_Node *Node, Uint64 Offset, Uint64 Length, void *Buffer)
+size_t SysFS_Comm_ReadFile(tVFS_Node *Node, off_t Offset, size_t Length, void *Buffer)
 {
 	if( Offset > Node->Size )	return -1;
 	if( Length > Node->Size )	Length = Node->Size;
